remove base module and move to utils
