--- modules/base/README.md
@@ -1,1 +1,0 @@
-# README: base
