--- modules/base/examples/example.cpp
@@ -1,12 +1,0 @@
-//=================================================================================================
-// Copyright (C) 2023-2024 HEPHAESTUS Contributors
-//=================================================================================================
-
-#include <cstdlib>
-
-//=================================================================================================
-auto main(int argc, const char* argv[]) -> int {
-  (void)argc;
-  (void)argv;
-  return EXIT_SUCCESS;
-}
