+++ modules/cli/src/program_options.cpp
@@ -6,7 +6,7 @@
 
 #include <fmt/core.h>
 
-#include "hephaestus/base/exception.h"
+#include "hephaestus/utils/exception.h"
 #include "hephaestus/utils/utils.h"
 
 namespace heph::cli {
