+++ modules/concurrency/CMakeLists.txt
@@ -4,7 +4,7 @@
 
 declare_module(
   NAME concurrency
-  DEPENDS_ON_MODULES base
+  DEPENDS_ON_MODULES utils
   DEPENDS_ON_EXTERNAL_PROJECTS absl fmt
 )
 
@@ -17,7 +17,7 @@ set(SOURCES src/spinner.cpp README.md include/hephaestus/concurrency/spinner.h)
 # library target
 define_module_library(
   NAME concurrency
-  PUBLIC_LINK_LIBS absl::log hephaestus::base fmt::fmt
+  PUBLIC_LINK_LIBS absl::log hephaestus::utils fmt::fmt
   PRIVATE_LINK_LIBS ""
   SOURCES ${SOURCES}
   PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include> $<INSTALL_INTERFACE:include>
