+++ modules/examples/CMakeLists.txt
@@ -1,6 +1,6 @@
 declare_module(
   NAME examples
-  DEPENDS_ON_MODULES bag base ipc serdes
+  DEPENDS_ON_MODULES bag utils ipc serdes
   DEPENDS_ON_EXTERNAL_PROJECTS absl Eigen3 fmt mcap Protobuf
   EXCLUDE_FROM_ALL
 )
@@ -35,7 +35,7 @@ set(TYPES_PROTOBUF_SOURCES
 # library target
 define_module_library(
   NAME examples_types_protobuf
-  PUBLIC_LINK_LIBS hephaestus::base hephaestus::serdes hephaestus::examples_types hephaestus::examples_types_proto
+  PUBLIC_LINK_LIBS hephaestus::utils hephaestus::serdes hephaestus::examples_types hephaestus::examples_types_proto
                    Eigen3::Eigen
   PRIVATE_LINK_LIBS ""
   SOURCES ${TYPES_PROTOBUF_SOURCES}
