+++ modules/ipc/include/hephaestus/ipc/zenoh/utils.h
@@ -10,8 +10,8 @@
 
 #include <zenohc.hxx>
 
-#include "hephaestus/base/exception.h"
 #include "hephaestus/ipc/common.h"
+#include "hephaestus/utils/exception.h"
 namespace heph::ipc::zenoh {
 
 /// We use single char key to reduce the overhead of the attachment.
