+++ modules/ipc/src/zenoh/utils.cpp
@@ -4,8 +4,8 @@
 
 #include "hephaestus/ipc/zenoh/utils.h"
 
-#include "hephaestus/base/exception.h"
 #include "hephaestus/ipc/common.h"
+#include "hephaestus/utils/exception.h"
 
 namespace heph::ipc::zenoh {
