+++ modules/utils/CMakeLists.txt
@@ -17,10 +17,12 @@ configure_file(src/version.in ${CMAKE_CURRENT_SOURCE_DIR}/src/version_impl.h @ON
 # library sources
 set(SOURCES
     src/concepts.cpp
-    src/version.cpp
+    src/exception.cpp
     src/utils.cpp
+    src/version.cpp
     README.md
     include/hephaestus/utils/concepts.h
+    include/hephaestus/utils/exception.h
     include/hephaestus/utils/utils.h
     include/hephaestus/utils/version.h
 )
